Use a Map and not an Object for Moo.js keyword transforms
We still need our own version of this code until the library is updated to
avoid eval, but the maintainers point out that using Maps is almost 10x
faster. I checked for our keywords and find over a 10x improvement.
Map looks to be widely supported in browsers so this ought to be simple.

Still worth watching no-context/moo#141 until the library is updated.